requirements.txt
robotpy==2026.2.1.1
robotpy-cli==2026.0.1
certifi
robotpy-rev
wpilib
phoenix6
robotpy-pathplannerlib
robotpy-commands-v2